refactor : dto 수정
- Admin dto 고유식별번호 int->long, 가격 int -> BigDecimal로 변경
- 요구사항에 따른 CommunityTopResponse dto 필드 추가
- mock 서버 더미데이터 수정

+++ apimocking-main/apimocking-boilerplate/src/main/java/com/grepp/spring/app/model/admin/dto/AdminStoreMenuResponse.java
@@ -1,6 +1,7 @@
 package com.grepp.spring.app.model.admin.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 
 @Schema(description = "관리자 가게 메뉴 응답 DTO")
 public record AdminStoreMenuResponse(
@@ -9,7 +10,7 @@ public record AdminStoreMenuResponse(
     String name,
 
     @Schema(description = "메뉴 가격", example = "4500")
-    int price
+    BigDecimal price
 
     ) {
